* commit 'ce4ad3946facb58ae843e6b1a6bc36c1e0aedee6':
  Fix UB (out-of-bounds-read) in PacBio::BAM::Complement
  Create proper artifact for PA

+++ scripts/ci/install.sh
@@ -5,6 +5,11 @@ set -vex
 # INSTALL #
 ###########
 
+if [[ ${_install_image} != true ]]; then
+  echo "Not installing image (branch: ${bamboo_planRepository_branchName}), returning."
+  return 0
+fi
+
 if [[ ${PREFIX_ARG} ]]; then
   ## Cleaning out old installation from /mnt/software
   rm -rf "${PREFIX_ARG}"/*
